+++ packages/api/graphql/mutations/process-attachment.graphql
@@ -1,0 +1,3 @@
+mutation ProcessAttachment($input: AnalyzeAttachmentInput!) {
+	mediaAnalyzeAttachment(input: $input)
+}
