--- packages/api/graphql/mutations/process-media.graphql
@@ -1,3 +1,0 @@
-mutation processMedia($input: AnalyzeAttachmentInput!) {
-  mediaAnalyzeAttachment(input: $input)
-}
