+++ packages/api/src/attachments/prepare.ts
@@ -16,7 +16,7 @@ export async function prepareAttachmentForUpload(
 	record: AttachableRecords,
 ) {
 	const isMultipart = data.size > MULTIPART_UPLOAD_CHUNK_SIZE;
-	const res = await this.uploadMedia({
+	const res = await this.UploadMedia({
 		input: {
 			byteSizeV2: data.size.toString(),
 			record,
