+++ packages/api/src/index.ts
@@ -1,15 +1,16 @@
-export { WhopApi, type WhopApiOptions } from "./api";
-export * as proto from "./codegen/proto";
+export type * from "@/codegen/generated-api";
+export * as proto from "@/codegen/proto";
+export { WhopClientSdk, type WhopClientSdkOptions } from "@/sdk.client";
+export { WhopServerSdk, type WhopServerSdkOptions } from "@/sdk.server";
 export {
 	getUserToken,
 	makeUserTokenVerifier,
 	verifyUserToken,
 	type UserTokenPayload,
 	type VerifyUserTokenOptions,
-} from "./verify-user-token";
-export { makeWebhookValidator, type WhopWebhookRequestBody } from "./webhooks";
+} from "@/verify-user-token";
+export { makeWebhookValidator, type WhopWebhookRequestBody } from "@/webhooks";
 export type {
 	ReceivableWebsocketMessage,
 	SendableWebsocketMessage,
-} from "./websockets";
-export type * from "./codegen/generated-api";
+} from "@/websockets";
